renamed histograms to histogram
